pyproject.toml
[project]
name = "cara-audio"
version = "0.1.0"
description = "GPU-accelerated TTS + STT audio backend for CARA"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { text = "MIT" }
authors = [{ name = "CARA Team" }]

dependencies = [
    # FastAPI + Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    
    # Settings & Validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    
    # PyTorch (CUDA 12.x)
    "torch>=2.5.0",
    "torchaudio>=2.5.0",
    "torchcodec>=0.2.1",
    
    # TTS - Chatterbox (MIT license, 23 languages incl. Italian)
    "chatterbox-tts",
    
    # STT - faster-whisper (CTranslate2)
    "faster-whisper>=1.0.0",
    
    # Audio processing
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    
    # Job queue
    "redis>=5.2.0",
    
    # Observability
    "prometheus-client>=0.21.0",
    
    # WebSocket support
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cara"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]
